Backend/requirements.txt
flask
flask-jwt-extended
pymongo
tensorflow
numpy
python-dotenv
pillow
werkzeug
flask-cors